Add files via upload
All of the files currently on my ionos server inside the Codetech folder are being uploaded. there will be 3 more folders to add, they are websites that I made in school at CSN.  All of my images are floating around in here with my index, css, and rm docs. I will need to implement better organization.  Ideally not only for this page, but id like to save my images in a layer closer to the top and make the pages go up to find and load them.  I don't want to save images repetitively throughout the site to be able to use them, that is redundant.